refactor php generator 1. resolve imports before init combinator 2. Integrate configurations and helper method 3. ObjectItem add subObject property 4. use `resolver` logic instead of `visitor` logic 5. deprecated some helper method 6. not supported `ast.predefined`
